app/requirements.txt
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-semantic-conventions==0.42b0
flask==3.0.2 